feat: Report builder
Adds a more flexible method to construct reports, which will allow
things such as providing an explicit backtrace, using a local eyre hook,
etc.
